requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==5.1.9
django-crum==0.7.9
django-widget-tweaks==1.5.1
pillow==12.1.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
